Remove utils for everything
