+++ AIAgent/common/file_system_utils.py
@@ -1,0 +1,17 @@
+import os
+from pathlib import Path
+
+
+def create_folders_if_necessary(paths: list[Path]) -> None:
+    for path in paths:
+        if not path.exists():
+            os.makedirs(path)
+
+
+def create_file(file: Path):
+    open(file, "w").close()
+
+
+def append_to_file(file: Path, s: str):
+    with open(file, "a") as file:
+        file.write(s)
